Fix tests failing due to new release of sklearn (0.21.0) - Enforce int type labels from stream generators - Reset sklearn version in requirements to >=0.20 - Closes #114

+++ src/skmultiflow/data/agrawal_generator.py
@@ -267,7 +267,7 @@ class AGRAWALGenerator(Stream):
             data[j, 9] = group
 
         self.current_sample_x = data[:, :self.n_features]
-        self.current_sample_y = data[:, self.n_features:].flatten()
+        self.current_sample_y = data[:, self.n_features:].flatten().astype(int)
 
         return self.current_sample_x, self.current_sample_y
